requirements.txt
# FastAPI y servidor
fastapi==0.120.0
uvicorn[standard]==0.38.0
pydantic==2.12.3

# Base de datos MongoDB
mongoengine==0.29.1
pymongo==3.12.0
dnspython==2.7.0

# Utilidades
python-dotenv==1.1.1
python-multipart==0.0.20
pyjwt==2.10.1

# HTTP requests
requests==2.32.5

# Testing
pytest==8.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
mongomock==4.1.2

